02_ingest/monthlyupdate/requirements.txt
google-cloud-storage==2.18.2
google-cloud-bigquery==3.26.0
functions-framework==3.*
google-cloud-logging